Move random.h from internal to bits directory
Issue #832
