+++ enclave/random.c
@@ -4,10 +4,10 @@
 #include "random.h"
 #include <mbedtls/ctr_drbg.h>
 #include <mbedtls/entropy.h>
+#include <openenclave/bits/random.h>
 #include <openenclave/enclave.h>
 #include <openenclave/internal/enclavelibc.h>
 #include <openenclave/internal/raise.h>
-#include <openenclave/internal/random.h>
 #include <openenclave/internal/thread.h>
 
 /*
