+++ include/openenclave/bits/random.h
