+++ tests/crypto/enclave/enc/enc.c
@@ -2,6 +2,7 @@
 // Licensed under the MIT License.
 
 #include <assert.h>
+#include <openenclave/bits/random.h>
 #include <openenclave/enclave.h>
 #include <openenclave/internal/cert.h>
 #include <openenclave/internal/ec.h>
@@ -9,7 +10,6 @@
 #include <openenclave/internal/hexdump.h>
 #include <openenclave/internal/malloc.h>
 #include <openenclave/internal/raise.h>
-#include <openenclave/internal/random.h>
 #include <openenclave/internal/rsa.h>
 #include <openenclave/internal/sha.h>
 #include <openenclave/internal/tests.h>
