+++ tests/crypto/host/main.c
@@ -2,11 +2,11 @@
 // Licensed under the MIT License.
 
 #include <assert.h>
+#include <openenclave/bits/random.h>
 #include <openenclave/internal/cert.h>
 #include <openenclave/internal/ec.h>
 #include <openenclave/internal/hexdump.h>
 #include <openenclave/internal/raise.h>
-#include <openenclave/internal/random.h>
 #include <openenclave/internal/rsa.h>
 #include <openenclave/internal/sha.h>
 #include <openenclave/internal/tests.h>
