+++ tests/crypto/random_tests.c
@@ -1,11 +1,12 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+#include <openenclave/bits/random.h>
+
 #if defined(OE_BUILD_ENCLAVE)
 #include <openenclave/enclave.h>
 #endif
 
-#include <openenclave/internal/random.h>
 #include <openenclave/internal/tests.h>
 #include <stdio.h>
 #include <string.h>
